+++ black_scholes/objects/gamma.c
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-int calculate_gamma( obj_loc_t obj_loc, int id )
+int calculate_gamma( obj_loc_t obj_loc, long id )
 {
     gamma_gamma = bs_delta_N_pd1 / ( stock_S * stock_trade_vRtT );
